@@ -1,6 +1,6 @@
 {
   "name": "ci",
-  "version": "0.9.0",
+  "version": "1.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -13,54 +13,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.1.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.1.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
     "@types/jquery": "^3.3.1",
     "@types/mathjax": "0.0.32",
     "angular-sortablejs": "^2.5.2",
     "angular2-fontawesome": "^5.2.1",
-    "copy-webpack-plugin": "4.5.0",
+    "copy-webpack-plugin": "4.5.1",
     "core-js": "^2.5.5",
     "expr-eval": "^1.2.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "mathjax": "^2.7.2",
-    "rxjs": "^5.5.0",
+    "rxjs": "^6.2.0",
     "sortablejs": "^1.7.0",
     "systemjs": "^0.21.3",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@types/node": "~9.4.6",
-    "codelyzer": "~4.3.0",
-    "ts-node": "~6.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "2.4.2",
-    "@types/jasmine": "~2.5.53",
+    "@angular/cli": "^6.0.3",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.2",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@types/node": "~10.1.2",
+    "codelyzer": "~4.3.0",
+    "jasmine": "~3.1.0",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.2",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2"
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "protractor": "~5.3.2",
+    "ts-node": "~6.0.1",
+    "tslint": "~5.10.0",
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.3"
   },
   "optionalDependencies": {
     "fsevents": "^1.2.2"
