@@ -1,7 +1,7 @@
 ﻿import { Injectable, Inject} from '@angular/core';
 import { Http, Headers, Response, RequestOptions } from '@angular/http';
 import { Observable } from 'rxjs';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 import { environment } from '../../environments/environment';
 
 @Injectable()
@@ -23,30 +23,32 @@ export class AuthenticationService {
         let options = new RequestOptions({ headers: headers }); // Create a request option
 
         return this.http.post(this.apiUrl+'/authenticate', request, options)
-            .map((response: Response) => {
-                // login successful if there's a jwt token in the response
-                let token = response.json() && response.json().message && 
-                    response.json().message.token;
-                if (token) {
-                    // set token property
-                    this.token = token;
+            .pipe(
+                map((response: Response) => {
+                    // login successful if there's a jwt token in the response
+                    let token = response.json() && response.json().message && 
+                        response.json().message.token;
+                    if (token) {
+                        // set token property
+                        this.token = token;    
 
-                    // store username and jwt token in local storage to keep user logged in between page refreshes
-                    localStorage.setItem('currentUser', 
-                        JSON.stringify({ username: username, token: token }));
-                    var question_num = 5;
-                    if (response.json().message && response.json()
-                        .message.question_num != undefined) {
-                        question_num = response.json().message.question_num;
+                        // store username and jwt token in local storage to keep user logged in between page refreshes
+                        localStorage.setItem('currentUser', 
+                            JSON.stringify({ username: username, token: token }));
+                        var question_num = 5;
+                        if (response.json().message && response.json()
+                            .message.question_num != undefined) {
+                            question_num = response.json().message.question_num;
+                        }
+                        localStorage.setItem('question_num', question_num+"");
+                        // return true to indicate successful login
+                        return true;
+                    } else {
+                        // return false to indicate failed login
+                        return false;
                     }
-                    localStorage.setItem('question_num', question_num+"");
-                    // return true to indicate successful login
-                    return true;
-                } else {
-                    // return false to indicate failed login
-                    return false;
-                }
-            });
+                })
+            );
     }
 
     register(username: string, email: string, password: string): Observable<boolean> {
@@ -56,12 +58,14 @@ export class AuthenticationService {
         let options = new RequestOptions({ headers: headers }); // Create a request option
 
         return this.http.post(this.apiUrl+'/register', request, options)
-            .map((response: Response) => {
-                // login successful if there's a jwt token in the response
-                let token = response.json() && response.json()
-                    .message && response.json().message.token;
-                return response.json();
-            });
+            .pipe(
+                map((response: Response) => {
+                    // login successful if there's a jwt token in the response
+                    let token = response.json() && response.json()
+                        .message && response.json().message.token;
+                    return response.json();
+                })
+            );
     }
 
     logout(): void {
@@ -73,13 +77,14 @@ export class AuthenticationService {
     sendPasswordResetEmail(email: string): Observable<boolean>  {
         let url = this.baseUrl+'/reset-password';
         let request = JSON.stringify({ email: email, url:  url });
-        console.log(request);
         let headers = new Headers({ 'Content-Type': 'application/json' }); // ... Set content type to JSON
         let options = new RequestOptions({ headers: headers }); // Create a request option
         return this.http.post(this.apiUrl+'/password-reset-email', request, options)
-            .map((response: Response) => {
-                return response.json();
-            });
+            .pipe(
+                map((response: Response) => {
+                    return response.json();
+                })
+            );
     }
 
     resetPassword(newPassword: string, confirmedPassword: string, token: string): Observable<boolean> {
@@ -88,8 +93,10 @@ export class AuthenticationService {
         let headers = new Headers({ 'Content-Type': 'application/json' }); // ... Set content type to JSON
         let options = new RequestOptions({ headers: headers }); // Create a request option
         return this.http.post(this.apiUrl+'/reset-password', request, options)
-            .map((response: Response) => {
-                return response.json();
-            });
+            .pipe(
+                map((response: Response) => {
+                    return response.json();
+                })
+            );
     }
 }
