@@ -1,7 +1,7 @@
 ﻿import { Injectable } from '@angular/core';
 import { Http, Headers, RequestOptions, Response } from '@angular/http';
 import { Observable } from 'rxjs';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 
 import { HttpService } from './http.service';
 
@@ -15,13 +15,17 @@ export class TrackingService {
         // notify api about lesson start
         if(lesson_id == -1) {
           /** TODO: change this HARDCODED value to testoutstart! */
-          return this.http.post('/', '')
           //return this.http.post('/testoutstart', '')
-            .map((response: Response) => response); 
+          return this.http.post('/', '')
+            .pipe(
+                map((response: Response) => response)
+            ); 
         }
         else {
         return this.http.post('/lesson/'+lesson_id+'/start', '')
-            .map((response: Response) => response);
+            .pipe(
+                map((response: Response) => response)
+            );
         }
     }
 
@@ -31,11 +35,15 @@ export class TrackingService {
             weak_questions: weak_questions });
         if(lesson_id == -1) {
           return this.http.post('/topic/'+topic_id+'/testoutdone', request)
-            .map((response: Response) => response); 
+            .pipe(
+                map((response: Response) => response)
+            ); 
         }
         else {
           return this.http.post('/lesson/'+lesson_id+'/done', request)
-            .map((response: Response) => response);
+            .pipe(
+                map((response: Response) => response)
+            );
         }
     }
 }
