@@ -1,10 +1,9 @@
 ﻿import { Injectable } from '@angular/core';
 import { Http, Headers, RequestOptions, Response } from '@angular/http';
 import { Observable } from 'rxjs';
-import 'rxjs/add/operator/map'
+import { map, catchError } from 'rxjs/operators';
 
 import { User } from '../_models/index';
-
 import { AuthenticationService } from './index';
 import { HttpService } from './http.service';
 
@@ -17,10 +16,13 @@ export class UserService {
 
     public getProfile() {
       return this.http.get('/profile')
-        .map((res:Response) => res)
-        .catch(error => {
-          throw Error(error);
-      });
+        .pipe(
+          map((res:Response) => res),
+          catchError(error => {
+            console.log(error);
+            throw Error(error);
+          })
+        );
     }
 
     public changeProfile(user: User) {
@@ -31,11 +33,13 @@ export class UserService {
       });
 
       return this.http.post('/profile', request)
-          .map((res: Response) => { })
-          .catch(error => {
+          .pipe(
+            map((res: Response) => { }),
+            catchError(error => {
               console.log(error);
               throw Error(error);
-      });
+            })
+          );
     }
 
     public changePassword(newPassword: string, 
@@ -47,10 +51,12 @@ export class UserService {
       });
 
       return this.http.post('/profile', request)
-        .map((res: Response) => { })
-        .catch(error => {
-          console.log(error);
-          throw Error(error);
-        });
+        .pipe(
+          map((res: Response) => { }),
+          catchError(error => {
+            console.log(error);
+            throw Error(error);
+          })
+        );
     }
 }
